Movement now controlled by keyboard.
Found a bug in the copypasta version of the clamping code where a
reference to 'p' was written as 'this'. Also, reused some code from
mobian.htm to tweak the existing handleKeyboard to allow the player
character 'p1' to move left and right.  Space and Shift can be used to
accelerate the character's vertical speed, but it doesn't yet work as a
jump.

Also, added a placeholder character image and a reference image for
future spritework.