Placing the MoveMarker works | Added draw method.
Had a bunch of issues because I didn't realize that to go from the
clientY (which is relative to upper-left corner of the browser's view)
to the mouse's position relative to the upperleft corner of the page
(have to add the document scrollTop to the clientY) and then making that
position relative to the canvas (subtract the canvas's offsetTop and
offsetLeft from corresponding mouse position) took a number of commands
I didn't realize existed until now. Progress! Next up, to make a floor
to stand on. Or to actually move towards the marker.